--- a/BedrockCore.cpp
+++ b/BedrockCore.cpp
@@ -82,7 +82,7 @@ bool BedrockCore::peekCommand(BedrockCommand& command) {
     return true;
 }
 
-bool BedrockCore::processCommand(BedrockCommand& command, bool beginTransaction) {
+bool BedrockCore::processCommand(BedrockCommand& command) {
     AutoTimer timer(command, BedrockCommand::PROCESS);
 
     // Convenience references to commonly used properties.
@@ -95,10 +95,11 @@ bool BedrockCore::processCommand(BedrockCommand& command, bool beginTransaction)
     // Keep track of whether we've modified the database and need to perform a `commit`.
     bool needsCommit = false;
     try {
-        if (beginTransaction) {
-            if (!_db.beginConcurrentTransaction()) {
-                throw "501 Failed to begin concurrent transaction";
-            }
+        // If a transaction was already begun in `peek`, then this is a no-op. We cal it here to support the case where
+        // peek created a httpsRequest and closed it's first transaction until the httpsRequest was complete, in which
+        // case we need to open a new transaction.
+        if (!_db.beginConcurrentTransaction()) {
+            throw "501 Failed to begin concurrent transaction";
         }
 
         // Loop across the plugins to see which wants to take this.