--- a/BedrockCore.cpp
+++ b/BedrockCore.cpp
@@ -95,10 +95,10 @@ bool BedrockCore::processCommand(BedrockCommand& command) {
     // Keep track of whether we've modified the database and need to perform a `commit`.
     bool needsCommit = false;
     try {
-        // If a transaction was already begun in `peek`, then this is a no-op. We cal it here to support the case where
+        // If a transaction was already begun in `peek`, then this is a no-op. We call it here to support the case where
         // peek created a httpsRequest and closed it's first transaction until the httpsRequest was complete, in which
         // case we need to open a new transaction.
-        if (!_db.beginConcurrentTransaction()) {
+        if (!_db.insideTransaction() && !_db.beginConcurrentTransaction()) {
             throw "501 Failed to begin concurrent transaction";
         }