--- a/BedrockCore.h
+++ b/BedrockCore.h
@@ -32,15 +32,18 @@ class BedrockCore : public SQLiteCore {
 
     // Process is the follow-up to `peek` if `peek` was insufficient to handle the command. It will only ever be called
     // on the master node, and should always be able to resolve the command completely. When a command is passed to
-    // `process`, the caller will have already begun a database transaction with either `BEGIN TRANSACTION` or
-    // `BEGIN CONCURRENT`, and it's up to `process` to add the rest of the transaction, without performing a `ROLLBACK`
-    // or `COMMIT`, which will be handled by the caller. It returns `true` if it has modified the database and requires
-    // the caller to perform a `commit`, and `false` if no changes requiring a `commit` have been made. Upon being
-    // returned `false`, the caller will perform a `ROLLBACK` of the empty transaction, and will not replicate the
-    // transaction to slave nodes. Upon being returned `true`, the caller will attempt to perform a `COMMIT` and
-    // replicate the transaction to slave nodes. It's allowable for this `COMMIT` to fail, in which case this command
-    // *will be passed to process again in the future to retry*.
-    bool processCommand(BedrockCommand& command);
+    // `process`, the caller will *usually* have already begun a database transaction with either `BEGIN TRANSACTION`
+    // or `BEGIN CONCURRENT`, and it's up to `process` to add the rest of the transaction, without performing a
+    // `ROLLBACK` or `COMMIT`, which will be handled by the caller. It returns `true` if it has modified the database
+    // and requires the caller to perform a `commit`, and `false` if no changes requiring a `commit` have been made.
+    // Upon being returned `false`, the caller will perform a `ROLLBACK` of the empty transaction, and will not
+    // replicate the transaction to slave nodes. Upon being returned `true`, the caller will attempt to perform a
+    // `COMMIT` and replicate the transaction to slave nodes. It's allowable for this `COMMIT` to fail, in which case
+    // this command *will be passed to process again in the future to retry*.
+    // The `beginTransaction` flag allows the caller to force this function to begin a transaction if it knows one has
+    // already been started. This is used to allow handling commands with httpsRequests without re-peeking them, which
+    // would potentially duplicate the httpsRequest.
+    bool processCommand(BedrockCommand& command, bool beginTransaction = false);
 
   private:
     void _handleCommandException(BedrockCommand& command, const string& e, bool wasProcessing);