There are two changes in this PR (you can look at them individual in each commit).

Change 1 is a correctness issue. It makes us include `peekCommand` in a transaction so that it's done atomically with `processCommand`. This involved a couple differences. We need to `rollback` after `peekCommand` now, because we've begun a transaction. And there is special handling for commands with `httpsRequests`, which won't get re-peeked, and thus are exempted from this single-transaction-per-command behavior.

Change 2 is a performance change, and while simpler, is the more exciting part of this PR. It does two things: First it moves the lock of `_syncThreadCommitMutex` to after `processCommand` in worker threads. Second, it logs the amount of time it takes the sync thread to acquire the same lock.

I did some analysis of the logs from our last test. I couldn't pinpoint exactly how much time we spent acquiring the lock in the sync thread, because we didn't have that info, but I timed the difference from when we dequeued a command until we started to commit a command, which includes waiting for this lock. Notably, this also includes `processCommand`, but that shouldn't affect the overall results.

Here's what I found in my sample data:

Before turning on multi-write:
19,862 commits totaling 50.01 seconds. Average 2.52ms per command.

While multi-write was enabled:
21048 commits totaling 419.20 seconds. Average 19.91ms per command.

After turning off multi-write:
15982 commits totaling 117.80 seconds. Average 7.37ms per command.

So, there's some pretty high variability there - before and after have average times that vary by a factor of 3x. But with multi-write enabled, the times are 2.5-10x slower than with it off. `processCommand` should take a significant amount of time, but it shouldn't change with the state of multi-write -- it runs in parallel on all threads.

So, the difference could be down to waiting for the lock, at an average of an extra 12-17ms per command. Is this a feasible time to wait? Well, if we have to wait for one worker thread to run `processCommand`, that seems to take 2.5-7ms. If we wait for two or three, this seems feasible. Even more so when you consider that retries due to conflicts will also fall into this category.

Note that the increase isn't uniform across commands. Most remain fast, but a few get markedly slower. A small number of commands can take over a second to process. If a worker runs one of these slow commands, before this change, the sync thread will block until the worker finishes before even starting, thus drastically increasing it's total time.

The actual time we see when we look at the entire lifetime of commands is mostly taken up waiting around in the sync thread queue, but that makes sense when the sync thread is just slow enough to fall behind. Each command takes longer to process, so the queue backs up, so new commands sit in it longer.

Please review carefully, all this code is fairly nuanced.