@tylerkaraszewski 2017-07-05
@quinthar - updated, with the trivial changes in one commit, and the change to not begin a transaction if one is already in progress in the other commit.